Fix id type mismatch and use i64/INT8 consistently

+++ backend/src/controllers/sessions.rs
@@ -34,7 +34,7 @@ pub async fn add(State(ctx): State<AppContext>, Json(params): Json<Params>) -> R
 }
 
 pub async fn update(
-    Path(id): Path<i32>,
+    Path(id): Path<i64>,
     State(ctx): State<AppContext>,
     Json(params): Json<Params>,
 ) -> Result<Response> {
@@ -45,12 +45,12 @@ pub async fn update(
     format::json(item)
 }
 
-pub async fn remove(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn remove(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     load_item(&ctx, id).await?.delete(&ctx.db).await?;
     format::empty()
 }
 
-pub async fn get_one(Path(id): Path<i32>, State(ctx): State<AppContext>) -> Result<Response> {
+pub async fn get_one(Path(id): Path<i64>, State(ctx): State<AppContext>) -> Result<Response> {
     format::json(load_item(&ctx, id).await?)
 }
 
@@ -65,7 +65,7 @@ pub fn routes() -> Routes {
         .add("/{id}", patch(update))
 }
 
-pub async fn load_item(ctx: &AppContext, id: i32) -> Result<Model> {
+pub async fn load_item(ctx: &AppContext, id: i64) -> Result<Model> {
     let item = Entity::find_by_id(id).one(&ctx.db).await?;
     item.ok_or_else(|| Error::NotFound)
 }
